services/producer/pyproject.toml
[project]
name = "producer"
version = "0.1.0"
description = "Producer pipeline to simulate events being produced to a Kafka topic."
authors = [
    {name = "Aaron Ginder QMK R",email = "aaronginder@hotmail.co.uk"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "apache-beam[gcp] (>=2.66.0,<3.0.0)",
    "ipykernel (>=6.30.0,<7.0.0)",
    "kafka-python (>=2.3.0,<3.0.0)",
]

[tool.poetry]
packages = [{include = "pipelines"}]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
